requirements.txt
# MCP Framework
mcp>=1.0.0

# Oracle Database Driver (최신)
oracledb>=2.0.0

# 데이터 처리
pandas>=2.0.0

# 보안 (암호화)
cryptography>=41.0.0

# 환경 변수
python-dotenv>=1.0.0

# 유틸리티
python-dateutil>=2.8.0
